Began specs for users controller.
I began implementing the skeleton specs for users controller. I've also
refactored several of the user's views to make them pass the specs.
Some specs still fail as of now.

Also included in this commit is a new database schema from the addition
of the user name index, some changes to my testing factories, and a
minor change on the routing file.