Removed devise in favor of google authenitaction.
This commit is a massive change.

First, I removed devise and replaced it with omniauth, using google as the
only accepted authentication provider. If you try to sign in without an
account, it gets your google openid, email address, and name, and fills
in a new user form with this name and email as the default values.

I've also added tests for the sign in process, replaced or edited the user
views, changed the user model and the user model specs, and began adaptations
of the users controller.