Fixed rsvps to only use one db table and added nested rsvp and link routes.
I changed the rsvp system to use only one table with a 'kind' column. The
three kinds used are 'host', 'attend', and 'maybe'. The models and model
specs have been updated to reflect this change. I added a method for each
kind to the user model to present all the events that they are hosting,
attending, or may be attending. Similar user returning methods were added
to the event model for each kind.

I also added event nested routes for links and rsvps, with appropriate specs.