Added user RSVP buttons to events show page.
I've added buttons to the events show page that allow users to RSVP to
the event, along with methods to User to support this. I also fixed
the events index view to show the venue name and link to the venue
show page, if a venue is associated with the event.