Added rank attribute to users.
I added the "rank" attribute to users, which will control updating and
deletion privileges. Rank images were added to the show and index views
for users, and emails were removed from all users views except for the
show and edit views for the currently signed in user.

I also made the layout link to 'events' actually link to the events index
action, and I refactored some of my tests using the spec_helper.