Reorganized views and edited the overall site layout.
I edited most of the views to present some of the data over on the side
bar. I also put the name of the object being viewed in an h1, with
a link to the edit page for that object right next to the object, if
appropriate.

I also fixed some bugs with the Events date and its time related scopes,
refactored some of the tests related to date scopes, and revised some other
controller tests to better match the desired view behavior.