Added user theme functionality into the MVC architecture.
Users now store their theme preferences in the database and can change
thier theme in their profile edit page. This commit contains the database
migration, user model changes, and view changes to support this.

In addition, I refactored the way themes are accessed, putting most of
theme logic that was in an initializer into the application controller.