Added tracked_rsvp email to UsersMailer.
This commit contains adds the tracked_rsvp email, which will be sent
when a user you are tracking rsvps to an event. It includes additons
to the spec helper to support specing the event date, and a refactoring
and expanding of the application helper to display event dates.

I refactored the mailer views to put salutations and signatures in a
layout, and to put displaying event information in a partial.

I also edited the development mailer interceptor to blank out cc and
bcc fields.