Refactored emails and added new user and rsvp emails.
I added a new user email, fired off by the create user action.

In addition, I've added an events email that gets triggered when a user
rsvps to an event. The email is sent to all other users who have rsvp'd
to the event and who have indicated in their preferences that the want to
be notified of rsvps.